+++ src/trx/game/objects/vars.c
@@ -97,6 +97,7 @@ const OBJECT_ID g_CreatureObjects[] = {
     O_NATLA,
     O_PATROL_DOG,
     O_PIERRE,
+    O_PRISONER,
     O_PUMA,
     O_PUNK_1,
     O_PUNK_2,